+++ app/Console/Commands/BarImportFromV3Export.php
@@ -15,7 +15,7 @@ class BarImportFromV3Export extends Command
      *
      * @var string
      */
-    protected $signature = 'bar:import-zip {filename?}';
+    protected $signature = 'bar:import-zip {filename?} {--password= : Set a new password for all imported users }';
 
     /**
      * The console command description.
@@ -75,7 +75,7 @@ class BarImportFromV3Export extends Command
         }
 
         try {
-            $this->exporter->process($zipFilePath);
+            $this->exporter->process($zipFilePath, $this->option('password') ?? null);
         } catch (Throwable $e) {
             $this->error($e->getMessage());
 
@@ -85,7 +85,7 @@ class BarImportFromV3Export extends Command
         Artisan::call('bar:refresh-search');
         Artisan::call('cache:clear');
 
-        $this->info('Importing finished!');
+        $this->output->success('Importing finished!');
 
         return Command::SUCCESS;
     }